* merge multi_multi @r:539 (and hence 0.2.x @r:538)  * added some missing svn:eol-style props

+++ source/game/facilities/pos_iterator.cpp
@@ -34,7 +34,7 @@ PosCircularIteratorFactory::PosCircularIteratorFactory(int maxRadius) :
 		for(int off = 0; off <= step; ++off) {
 			p->step = step;
 			p->off = off;
-			p->dist = sqrtf(static_cast<float>(step * step + off * off));
+			p->dist = int(sqrtf(float(step * step + off * off)));
 			++p;
 		}
 	}
